Move Random Battles files to a new folder
