--- data/mods/gen7pokebilities/random-teams.ts
@@ -1,5 +1,0 @@
-import RandomGen7Teams from '../gen7/random-teams';
-
-export class RandomGen7PokebilitiesTeams extends RandomGen7Teams {}
-
-export default RandomGen7PokebilitiesTeams;
