--- data/mods/partnersincrime/random-teams.ts
@@ -1,5 +1,0 @@
-import RandomTeams from '../../random-teams';
-
-export class RandomPartnersInCrimeTeams extends RandomTeams {}
-
-export default RandomPartnersInCrimeTeams;
