+++ data/random-battles/gen5-teams.ts
@@ -1,20 +1,16 @@
-import {MoveCounter, TeamData} from '../gen8/random-teams';
-import RandomGen7Teams, {BattleFactorySpecies, ZeroAttackHPIVs} from '../gen7/random-teams';
-import {PRNG, PRNGSeed} from '../../../sim/prng';
-import {Utils} from '../../../lib';
-import {toID} from '../../../sim/dex';
+import RandomGen6Teams from './gen6-teams';
+import {Utils} from '../../lib';
+import {toID} from '../../sim/dex';
+import {PRNG} from '../../sim';
+import {MoveCounter} from './gen8-teams';
 
 // Moves that restore HP:
 const RECOVERY_MOVES = [
-	'healorder', 'milkdrink', 'moonlight', 'morningsun', 'recover', 'recycle', 'roost', 'slackoff', 'softboiled', 'synthesis',
+	'healorder', 'milkdrink', 'moonlight', 'morningsun', 'recover', 'roost', 'slackoff', 'softboiled', 'synthesis',
 ];
 // Moves that boost Attack:
 const PHYSICAL_SETUP = [
-	'bellydrum', 'bulkup', 'coil', 'curse', 'dragondance', 'honeclaws', 'howl', 'meditate', 'poweruppunch', 'screech', 'swordsdance',
-];
-// Moves which boost Special Attack:
-const SPECIAL_SETUP = [
-	'calmmind', 'chargebeam', 'geomancy', 'nastyplot', 'quiverdance', 'tailglow',
+	'bellydrum', 'bulkup', 'coil', 'curse', 'dragondance', 'honeclaws', 'howl', 'meditate', 'screech', 'swordsdance',
 ];
 // Some moves that only boost Speed:
 const SPEED_SETUP = [
@@ -23,27 +19,23 @@ const SPEED_SETUP = [
 // Conglomerate for ease of access
 const SETUP = [
 	'acidarmor', 'agility', 'autotomize', 'bellydrum', 'bulkup', 'calmmind', 'coil', 'curse', 'dragondance', 'flamecharge',
-	'focusenergy', 'geomancy', 'growth', 'honeclaws', 'howl', 'irondefense', 'meditate', 'nastyplot', 'poweruppunch',
-	'quiverdance', 'raindance', 'rockpolish', 'shellsmash', 'shiftgear', 'swordsdance', 'tailglow', 'workup',
+	'growth', 'honeclaws', 'howl', 'irondefense', 'meditate', 'nastyplot', 'quiverdance', 'raindance', 'rockpolish',
+	'shellsmash', 'shiftgear', 'sunnyday', 'swordsdance', 'tailglow', 'workup',
 ];
 // Moves that shouldn't be the only STAB moves:
 const NO_STAB = [
-	'aquajet', 'bulletpunch', 'clearsmog', 'dragontail', 'eruption', 'explosion', 'fakeout', 'flamecharge',
-	'futuresight', 'iceshard', 'icywind', 'incinerate', 'machpunch', 'nuzzle', 'pluck', 'poweruppunch', 'pursuit',
-	'quickattack', 'rapidspin', 'reversal', 'selfdestruct', 'shadowsneak', 'skyattack', 'skydrop', 'snarl',
-	'suckerpunch', 'uturn', 'watershuriken', 'vacuumwave', 'voltswitch', 'waterspout',
+	'aquajet', 'bulletpunch', 'chatter', 'clearsmog', 'dragontail', 'eruption', 'explosion', 'fakeout', 'flamecharge',
+	'futuresight', 'iceshard', 'icywind', 'incinerate', 'knockoff', 'machpunch', 'pluck', 'pursuit', 'quickattack',
+	'rapidspin', 'reversal', 'selfdestruct', 'shadowsneak', 'skyattack', 'skydrop', 'snarl', 'suckerpunch',
+	'uturn', 'vacuumwave', 'voltswitch', 'waterspout',
 ];
 // Hazard-setting moves
 const HAZARDS = [
-	'spikes', 'stealthrock', 'stickyweb', 'toxicspikes',
-];
-// Protect and its variants
-const PROTECT_MOVES = [
-	'kingsshield', 'protect', 'spikyshield',
+	'spikes', 'stealthrock', 'toxicspikes',
 ];
 // Moves that switch the user out
 const PIVOT_MOVES = [
-	'partingshot', 'uturn', 'voltswitch',
+	'uturn', 'voltswitch',
 ];
 
 // Moves that should be paired together when possible
@@ -51,39 +43,33 @@ const MOVE_PAIRS = [
 	['lightscreen', 'reflect'],
 	['sleeptalk', 'rest'],
 	['protect', 'wish'],
-	['spikyshield', 'wish'],
 	['leechseed', 'substitute'],
-	['perishsong', 'protect'],
-	['solarbeam', 'sunnyday'],
 ];
 
 /** Pokemon who always want priority STAB, and are fine with it as its only STAB move of that type */
 const PRIORITY_POKEMON = [
-	'aegislash', 'banette', 'breloom', 'cacturne', 'doublade', 'dusknoir', 'honchkrow', 'scizor', 'scizormega', 'shedinja',
+	'bisharp', 'breloom', 'cacturne', 'dusknoir', 'honchkrow', 'scizor', 'shedinja', 'shiftry',
 ];
 
-export class RandomGen6Teams extends RandomGen7Teams {
-	randomSets: {[species: string]: RandomTeamsTypes.RandomSpeciesData} = require('./random-sets.json');
+export class RandomGen5Teams extends RandomGen6Teams {
+	randomSets: {[species: string]: RandomTeamsTypes.RandomSpeciesData} = require('./gen5-sets.json');
 
-	constructor(format: Format | string, prng: PRNG | PRNGSeed | null) {
+	constructor(format: string | Format, prng: PRNG | PRNGSeed | null) {
 		super(format, prng);
-
 		this.noStab = NO_STAB;
 		this.priorityPokemon = PRIORITY_POKEMON;
 
 		this.moveEnforcementCheckers = {
 			Bug: (movePool, moves, abilities, types, counter) => (
-				['megahorn', 'pinmissile'].some(m => movePool.includes(m)) ||
-				!counter.get('Bug') && abilities.has('Tinted Lens')
+				!counter.get('Bug') && (movePool.includes('megahorn') || abilities.has('Tinted Lens'))
 			),
 			Dark: (movePool, moves, abilities, types, counter) => !counter.get('Dark'),
 			Dragon: (movePool, moves, abilities, types, counter) => !counter.get('Dragon'),
 			Electric: (movePool, moves, abilities, types, counter) => !counter.get('Electric'),
-			Fairy: (movePool, moves, abilities, types, counter) => !counter.get('Fairy'),
 			Fighting: (movePool, moves, abilities, types, counter) => !counter.get('Fighting'),
 			Fire: (movePool, moves, abilities, types, counter) => !counter.get('Fire'),
 			Flying: (movePool, moves, abilities, types, counter, species) => (
-				!counter.get('Flying') && !['aerodactylmega', 'charizardmegay', 'mantine', 'murkrow'].includes(species.id) &&
+				!counter.get('Flying') && !['mantine', 'murkrow'].includes(species.id) &&
 				!movePool.includes('hiddenpowerflying')
 			),
 			Ghost: (movePool, moves, abilities, types, counter) => !counter.get('Ghost'),
@@ -91,12 +77,10 @@ export class RandomGen6Teams extends RandomGen7Teams {
 				!counter.get('Grass') && (species.baseStats.atk >= 100 || movePool.includes('leafstorm'))
 			),
 			Ground: (movePool, moves, abilities, types, counter) => !counter.get('Ground'),
-			Ice: (movePool, moves, abilities, types, counter) => (
-				!counter.get('Ice') || (!moves.has('blizzard') && movePool.includes('freezedry')) ||
-				abilities.has('Refrigerate') && (movePool.includes('return') || movePool.includes('hypervoice'))
+			Ice: (movePool, moves, abilities, types, counter) => !counter.get('Ice'),
+			Poison: (movePool, moves, abilities, types, counter) => (
+				!counter.get('Poison') && (types.has('Grass') || types.has('Ground'))
 			),
-			Normal: movePool => movePool.includes('boomburst'),
-			Poison: (movePool, moves, abilities, types, counter) => !counter.get('Poison'),
 			Psychic: (movePool, moves, abilities, types, counter) => (
 				!counter.get('Psychic') && (types.has('Fighting') || movePool.includes('calmmind'))
 			),
@@ -104,7 +88,7 @@ export class RandomGen6Teams extends RandomGen7Teams {
 				!counter.get('Rock') && (species.baseStats.atk >= 95 || abilities.has('Rock Head'))
 			),
 			Steel: (movePool, moves, abilities, types, counter, species) => (
-				!counter.get('Steel') && species.baseStats.atk >= 100
+				!counter.get('Steel') && ['aggron', 'metagross'].includes(species.id)
 			),
 			Water: (movePool, moves, abilities, types, counter) => !counter.get('Water'),
 		};
@@ -172,16 +156,11 @@ export class RandomGen6Teams extends RandomGen7Teams {
 			if (movePool.includes('lightscreen')) this.fastPop(movePool, movePool.indexOf('lightscreen'));
 			if (moves.size + movePool.length <= this.maxMoveCount) return;
 		}
-		if (teamDetails.stickyWeb) {
-			if (movePool.includes('stickyweb')) this.fastPop(movePool, movePool.indexOf('stickyweb'));
-			if (moves.size + movePool.length <= this.maxMoveCount) return;
-		}
 		if (teamDetails.stealthRock) {
 			if (movePool.includes('stealthrock')) this.fastPop(movePool, movePool.indexOf('stealthrock'));
 			if (moves.size + movePool.length <= this.maxMoveCount) return;
 		}
-		if (teamDetails.defog || teamDetails.rapidSpin) {
-			if (movePool.includes('defog')) this.fastPop(movePool, movePool.indexOf('defog'));
+		if (teamDetails.rapidSpin) {
 			if (movePool.includes('rapidspin')) this.fastPop(movePool, movePool.indexOf('rapidspin'));
 			if (moves.size + movePool.length <= this.maxMoveCount) return;
 		}
@@ -195,56 +174,47 @@ export class RandomGen6Teams extends RandomGen7Teams {
 		}
 
 		// Develop additional move lists
-		const badWithSetup = ['defog', 'dragontail', 'haze', 'healbell', 'nuzzle', 'pursuit', 'rapidspin', 'toxic'];
+		const badWithSetup = ['healbell', 'pursuit', 'toxic'];
+		// Nature Power is Earthquake this gen
 		const statusMoves = this.dex.moves.all()
-			.filter(move => move.category === 'Status')
+			.filter(move => move.category === 'Status' && move.id !== 'naturepower')
 			.map(move => move.id);
 
 		// General incompatibilities
 		const incompatiblePairs = [
 			// These moves don't mesh well with other aspects of the set
-			[statusMoves, ['healingwish', 'memento', 'switcheroo', 'trick']],
+			[statusMoves, ['healingwish', 'switcheroo', 'trick']],
 			[SETUP, PIVOT_MOVES],
 			[SETUP, HAZARDS],
 			[SETUP, badWithSetup],
 			[PHYSICAL_SETUP, PHYSICAL_SETUP],
-			[SPEED_SETUP, 'quickattack'],
-			['defog', HAZARDS],
 			[['fakeout', 'uturn'], ['switcheroo', 'trick']],
 			['substitute', PIVOT_MOVES],
-			['leechseed', 'dragontail'],
 			['rest', 'substitute'],
-			[PHYSICAL_SETUP, 'dracometeor'],
-			[SPECIAL_SETUP, 'knockoff'],
 
 			// These attacks are redundant with each other
 			['psychic', 'psyshock'],
-			['scald', ['hydropump', 'originpulse', 'waterpulse']],
-			['return', ['bodyslam', 'doubleedge']],
-			[['fierydance', 'lavaplume'], ['fireblast', 'magmastorm']],
-			[['flamethrower', 'flareblitz'], ['fireblast', 'overheat']],
-			['hornleech', 'woodhammer'],
+			[['scald', 'surf'], 'hydropump'],
+			[['bodyslam', 'return'], ['bodyslam', 'doubleedge']],
 			[['gigadrain', 'leafstorm'], ['leafstorm', 'petaldance', 'powerwhip']],
-			['wildcharge', 'thunderbolt'],
-			['gunkshot', 'poisonjab'],
 			[['drainpunch', 'focusblast'], ['closecombat', 'highjumpkick', 'superpower']],
-			['stoneedge', 'headsmash'],
-			['dracometeor', 'dragonpulse'],
-			['dragonclaw', 'outrage'],
-			['knockoff', ['darkpulse', 'foulplay']],
-
-			// Status move incompatibilities
-			['toxic', 'toxicspikes'],
-			['taunt', 'disable'],
-			['defog', ['leechseed', 'substitute']],
+			['payback', 'pursuit'],
 
 			// Assorted hardcodes go here:
+			// Zebstrika
+			['wildcharge', 'thunderbolt'],
+			// Manectric
+			['flamethrower', 'overheat'],
+			// Meganium
+			['leechseed', 'dragontail'],
+			// Volcarona and Heatran
+			[['fierydance', 'lavaplume'], 'fireblast'],
+			// Walrein
+			['encore', 'roar'],
 			// Lunatone
 			['moonlight', 'rockpolish'],
 			// Smeargle
-			['destinybond', 'whirlwind'],
-			// Liepard
-			['copycat', 'uturn'],
+			['memento', 'whirlwind'],
 			// Seviper
 			['switcheroo', 'suckerpunch'],
 			// Jirachi
@@ -253,21 +223,14 @@ export class RandomGen6Teams extends RandomGen7Teams {
 
 		for (const pair of incompatiblePairs) this.incompatibleMoves(moves, movePool, pair[0], pair[1]);
 
-		if (!types.includes('Dark') && preferredType !== 'Dark') {
-			this.incompatibleMoves(moves, movePool, 'knockoff', ['pursuit', 'suckerpunch']);
-		}
+		if (species.id === 'dugtrio') this.incompatibleMoves(moves, movePool, statusMoves, 'memento');
 
-		const statusInflictingMoves = ['thunderwave', 'toxic', 'willowisp', 'yawn'];
+		const statusInflictingMoves = ['stunspore', 'thunderwave', 'toxic', 'willowisp', 'yawn'];
 		if (!abilities.has('Prankster') && role !== 'Staller') {
 			this.incompatibleMoves(moves, movePool, statusInflictingMoves, statusInflictingMoves);
 		}
 
 		if (abilities.has('Guts')) this.incompatibleMoves(moves, movePool, 'protect', 'swordsdance');
-
-		// Force Protect and U-turn on Beedrill-Mega
-		if (species.id === 'beedrillmega') {
-			this.incompatibleMoves(moves, movePool, 'drillrun', 'knockoff');
-		}
 	}
 
 	// Generate random moveset for a given species, role, preferred type.
@@ -319,8 +282,8 @@ export class RandomGen6Teams extends RandomGen7Teams {
 				movePool, preferredType, role);
 		}
 
-		// Enforce Blizzard, Seismic Toss, Spore, and Sticky Web
-		for (const moveid of ['blizzard', 'seismictoss', 'spore', 'stickyweb']) {
+		// Enforce Seismic Toss and Spore
+		for (const moveid of ['seismictoss', 'spore']) {
 			if (movePool.includes(moveid)) {
 				counter = this.addMove(moveid, moves, types, abilities, teamDetails, species, isLead,
 					movePool, preferredType, role);
@@ -333,26 +296,16 @@ export class RandomGen6Teams extends RandomGen7Teams {
 				movePool, preferredType, role);
 		}
 
-		// Enforce Shadow Sneak on Kecleon
-		if (movePool.includes('shadowsneak') && species.id === 'kecleon') {
-			counter = this.addMove('shadowsneak', moves, types, abilities, teamDetails, species, isLead,
-				movePool, preferredType, role);
-		}
-
-		// Enforce hazard removal on Bulky Support if the team doesn't already have it
-		if (role === 'Bulky Support' && !teamDetails.defog && !teamDetails.rapidSpin) {
+		// Enforce hazard removal on Bulky Support and Spinner if the team doesn't already have it
+		if (['Bulky Support', 'Spinner'].includes(role) && !teamDetails.rapidSpin) {
 			if (movePool.includes('rapidspin')) {
 				counter = this.addMove('rapidspin', moves, types, abilities, teamDetails, species, isLead,
-					movePool, preferredType, role);
-			}
-			if (movePool.includes('defog')) {
-				counter = this.addMove('defog', moves, types, abilities, teamDetails, species, isLead,
 					movePool, preferredType, role);
 			}
 		}
 
 		// Enforce STAB priority
-		if (['Bulky Attacker', 'Bulky Setup', 'Wallbreaker'].includes(role) || this.priorityPokemon.includes(species.id)) {
+		if (['Bulky Attacker', 'Bulky Setup'].includes(role) || this.priorityPokemon.includes(species.id)) {
 			const priorityMoves = [];
 			for (const moveid of movePool) {
 				const move = this.dex.moves.get(moveid);
@@ -428,7 +381,7 @@ export class RandomGen6Teams extends RandomGen7Teams {
 		}
 
 		// Enforce recovery
-		if (['Bulky Support', 'Bulky Attacker', 'Bulky Setup', 'Staller'].includes(role)) {
+		if (['Bulky Support', 'Bulky Attacker', 'Bulky Setup', 'Spinner', 'Staller'].includes(role)) {
 			const recoveryMoves = movePool.filter(moveid => RECOVERY_MOVES.includes(moveid));
 			if (recoveryMoves.length) {
 				const moveid = this.sample(recoveryMoves);
@@ -439,7 +392,7 @@ export class RandomGen6Teams extends RandomGen7Teams {
 
 		// Enforce Staller moves
 		if (role === 'Staller') {
-			const enforcedMoves = [...PROTECT_MOVES, 'toxic', 'wish'];
+			const enforcedMoves = ['protect', 'toxic', 'wish'];
 			for (const move of enforcedMoves) {
 				if (movePool.includes(move)) {
 					counter = this.addMove(move, moves, types, abilities, teamDetails, species, isLead,
@@ -450,11 +403,20 @@ export class RandomGen6Teams extends RandomGen7Teams {
 
 		// Enforce setup
 		if (role.includes('Setup')) {
-			const setupMoves = movePool.filter(moveid => SETUP.includes(moveid));
-			if (setupMoves.length) {
-				const moveid = this.sample(setupMoves);
+			// First, try to add a non-Speed setup move
+			const nonSpeedSetupMoves = movePool.filter(moveid => SETUP.includes(moveid) && !SPEED_SETUP.includes(moveid));
+			if (nonSpeedSetupMoves.length) {
+				const moveid = this.sample(nonSpeedSetupMoves);
 				counter = this.addMove(moveid, moves, types, abilities, teamDetails, species, isLead,
 					movePool, preferredType, role);
+			} else {
+				// No non-Speed setup moves, so add any (Speed) setup move
+				const setupMoves = movePool.filter(moveid => SETUP.includes(moveid));
+				if (setupMoves.length) {
+					const moveid = this.sample(setupMoves);
+					counter = this.addMove(moveid, moves, types, abilities, teamDetails, species, isLead,
+						movePool, preferredType, role);
+				}
 			}
 		}
 
@@ -527,29 +489,25 @@ export class RandomGen6Teams extends RandomGen7Teams {
 		role: RandomTeamsTypes.Role
 	): boolean {
 		switch (ability) {
-		case 'Flare Boost': case 'Gluttony': case 'Harvest': case 'Ice Body': case 'Magician':
-		case 'Moody': case 'Pressure': case 'Sand Veil': case 'Sniper': case 'Snow Cloak': case 'Steadfast':
+		case 'Flare Boost': case 'Gluttony': case 'Ice Body': case 'Infiltrator': case 'Moody': case 'Pickpocket':
+		case 'Pressure': case 'Sand Veil': case 'Sniper': case 'Snow Cloak': case 'Steadfast': case 'Unburden':
 			return true;
-		case 'Aerilate': case 'Pixilate': case 'Refrigerate':
-			return ['doubleedge', 'hypervoice', 'return'].every(m => !moves.has(m));
 		case 'Chlorophyll':
 			// Petal Dance is for Lilligant
 			return (
 				species.baseStats.spe > 100 || moves.has('petaldance') ||
 				(!moves.has('sunnyday') && !teamDetails.sun)
 			);
-		case 'Competitive':
-			return !counter.get('Special');
 		case 'Compound Eyes': case 'No Guard':
 			return !counter.get('inaccurate');
-		case 'Contrary': case 'Skill Link': case 'Strong Jaw':
+		case 'Contrary': case 'Skill Link':
 			return !counter.get(toID(ability));
 		case 'Defiant': case 'Justified':
 			return !counter.get('Physical');
 		case 'Guts':
 			return (!moves.has('facade') && !moves.has('sleeptalk'));
 		case 'Hustle':
-			return counter.get('Physical') < 2;
+			return (counter.get('Physical') < 2 || species.id === 'delibird');
 		case 'Hydration': case 'Rain Dish': case 'Swift Swim':
 			return (
 				species.baseStats.spe > 100 || !moves.has('raindance') && !teamDetails.rain ||
@@ -565,52 +523,51 @@ export class RandomGen6Teams extends RandomGen7Teams {
 		case 'Magic Guard': case 'Speed Boost':
 			return (abilities.has('Tinted Lens') && role === 'Wallbreaker');
 		case 'Mold Breaker':
-			return (species.baseSpecies === 'Basculin' || species.id === 'pangoro' || abilities.has('Sheer Force'));
+			return (species.baseSpecies === 'Basculin' || species.id === 'rampardos');
 		case 'Moxie':
-			return (!counter.get('Physical') || moves.has('stealthrock') || (!!species.isMega && abilities.has('Intimidate')));
-		case 'Oblivious': case 'Prankster':
-			return (!counter.get('Status') || (species.id === 'tornadus' && moves.has('bulkup')));
-		case 'Overcoat':
-			return types.has('Grass');
+			return (!counter.get('Physical') || moves.has('stealthrock'));
 		case 'Overgrow':
 			return !counter.get('Grass');
+		case 'Prankster':
+			return (!counter.get('Status') || (species.id === 'tornadus' && moves.has('bulkup')));
+		case 'Poison Heal':
+			return (species.id === 'breloom' && role === 'Fast Attacker');
 		case 'Shed Skin':
 			return !moves.has('rest');
 		case 'Synchronize':
-			return (counter.get('Status') < 2 || !!counter.get('recoil') || !!species.isMega);
+			return (counter.get('Status') < 2 || !!counter.get('recoil'));
 		case 'Regenerator':
-			return species.id === 'mienshao' || species.id === 'reuniclus';
+			return ((species.id === 'mienshao' && role !== 'Fast Attacker') || species.id === 'reuniclus');
 		case 'Reckless': case 'Rock Head':
-			return (!counter.get('recoil') || !!species.isMega);
+			return !counter.get('recoil');
 		case 'Sand Force': case 'Sand Rush':
 			return !teamDetails.sand;
-		case 'Scrappy':
-			return !types.has('Normal');
 		case 'Serene Grace':
 			return !counter.get('serenegrace');
 		case 'Sheer Force':
-			return (!counter.get('sheerforce') || moves.has('doubleedge') || abilities.has('Guts') || !!species.isMega);
+			return (!counter.get('sheerforce') || moves.has('doubleedge') || abilities.has('Guts'));
 		case 'Simple':
 			return !counter.get('setup');
-		case 'Snow Warning':
-			// Aurorus
-			return moves.has('hypervoice');
 		case 'Solar Power':
-			return (!counter.get('Special') || !teamDetails.sun || !!species.isMega);
+			return (!counter.get('Special') || !teamDetails.sun);
+		case 'Sticky Hold':
+			return species.id !== 'accelgor';
 		case 'Sturdy':
 			return (!!counter.get('recoil') && !counter.get('recovery') || species.id === 'steelix' && !!counter.get('sheerforce'));
 		case 'Swarm':
-			return ((!counter.get('Bug') && !moves.has('uturn')) || !!species.isMega);
+			return !counter.get('Bug') && !moves.has('uturn');
 		case 'Technician':
-			return (!counter.get('technician') || moves.has('tailslap') || !!species.isMega);
+			return (!counter.get('technician') || moves.has('tailslap'));
 		case 'Tinted Lens':
-			return (['illumise', 'sigilyph', 'yanmega'].some(m => species.id === (m)) && role !== 'Wallbreaker');
+			// Night Shade part is for Noctowl
+			return (
+				moves.has('nightshade') ||
+				['illumise', 'sigilyph', 'yanmega'].some(m => species.id === (m)) && role !== 'Wallbreaker'
+			);
 		case 'Torrent':
-			return (!counter.get('Water') || !!species.isMega);
+			return !counter.get('Water');
 		case 'Unaware':
-			return (role !== 'Bulky Support' && role !== 'Staller');
-		case 'Unburden':
-			return (!!species.isMega || !counter.get('setup') && !moves.has('acrobatics'));
+			return ((role !== 'Bulky Attacker' && role !== 'Bulky Setup') || species.id === 'swoobat');
 		case 'Water Absorb':
 			return moves.has('raindance') || ['Drizzle', 'Unaware', 'Volt Absorb'].some(abil => abilities.has(abil));
 		}
@@ -630,9 +587,6 @@ export class RandomGen6Teams extends RandomGen7Teams {
 		preferredType: string,
 		role: RandomTeamsTypes.Role,
 	): string {
-		if (species.battleOnly && !species.requiredAbility) {
-			abilities = new Set(Object.values(this.dex.species.get(species.battleOnly as string).abilities));
-		}
 		const abilityData = Array.from(abilities).map(a => this.dex.abilities.get(a));
 		Utils.sortBy(abilityData, abil => -abil.rating);
 
@@ -644,35 +598,26 @@ export class RandomGen6Teams extends RandomGen7Teams {
 			!abilities.has('Quick Feet') &&
 			(moves.has('facade') || (moves.has('sleeptalk') && moves.has('rest')))
 		) return 'Guts';
-
 		if (species.id === 'starmie') return role === 'Wallbreaker' ? 'Analytic' : 'Natural Cure';
 		if (species.id === 'beheeyem') return 'Analytic';
 		if (species.id === 'ninetales') return 'Drought';
-		if (species.baseSpecies === 'Gourgeist') return 'Frisk';
-		if (species.id === 'pinsirmega') return 'Hyper Cutter';
-		if (species.id === 'ninjask' || species.id === 'seviper') return 'Infiltrator';
 		if (species.id === 'gligar') return 'Immunity';
 		if (species.id === 'arcanine' || species.id === 'stantler') return 'Intimidate';
-		if (species.id === 'lucariomega') return 'Justified';
-		if (species.id === 'persian' && !counter.get('technician')) return 'Limber';
-		if (species.baseSpecies === 'Altaria') return 'Natural Cure';
+		if (species.id === 'altaria') return 'Natural Cure';
+		if (species.id === 'mandibuzz') return 'Overcoat';
 		// If Ambipom doesn't qualify for Technician, Skill Link is useless on it
 		if (species.id === 'ambipom' && !counter.get('technician')) return 'Pickup';
-		if (species.id === 'muk') return 'Poison Touch';
-		if (['dusknoir', 'vespiquen'].includes(species.id)) return 'Pressure';
-		if (species.id === 'druddigon' && role === 'Bulky Support') return 'Rough Skin';
+		if (['spiritomb', 'vespiquen', 'weavile'].includes(species.id)) return 'Pressure';
+		if (species.id === 'druddigon') return 'Rough Skin';
 		if (species.id === 'zebstrika') return moves.has('wildcharge') ? 'Sap Sipper' : 'Lightning Rod';
-		if (species.id === 'stoutland' || species.id === 'pangoro' && !counter.get('ironfist')) return 'Scrappy';
+		if (species.id === 'stoutland') return 'Scrappy';
 		if (species.id === 'octillery') return 'Sniper';
 		if (species.id === 'stunfisk') return 'Static';
-		if (species.id === 'breloom') return 'Technician';
 		if (species.id === 'zangoose') return 'Toxic Boost';
 
-		if (abilities.has('Harvest') && (role === 'Bulky Support' || role === 'Staller')) return 'Harvest';
-		if (abilities.has('Regenerator') && role === 'AV Pivot') return 'Regenerator';
+		if (abilities.has('Harvest')) return 'Harvest';
 		if (abilities.has('Shed Skin') && moves.has('rest') && !moves.has('sleeptalk')) return 'Shed Skin';
-		if (abilities.has('Sniper') && moves.has('focusenergy')) return 'Sniper';
-		if (abilities.has('Unburden') && ['acrobatics', 'bellydrum'].some(m => moves.has(m))) return 'Unburden';
+		if (abilities.has('Unburden') && ['acrobatics', 'closecombat'].some(m => moves.has(m))) return 'Unburden';
 
 		let abilityAllowed: Ability[] = [];
 		// Obtain a list of abilities that are allowed (not culled)
@@ -731,24 +676,23 @@ export class RandomGen6Teams extends RandomGen7Teams {
 		role: RandomTeamsTypes.Role,
 	): string | undefined {
 		if (species.requiredItems) return this.sample(species.requiredItems);
-		if (role === 'AV Pivot') return 'Assault Vest';
-		if (species.name === 'Farfetch\u2019d') return 'Stick';
-		if (species.name === 'Latias' || species.name === 'Latios') return 'Soul Dew';
-		if (species.name === 'Marowak') return 'Thick Club';
-		if (species.name === 'Pikachu') return 'Light Ball';
-		if (species.name === 'Shedinja' || species.name === 'Smeargle') return 'Focus Sash';
-		if (species.name === 'Talonflame') return 'Sharp Beak';
-		if (species.name === 'Unfezant' || moves.has('focusenergy')) return 'Scope Lens';
-		if (species.name === 'Unown') return 'Choice Specs';
-		if (species.name === 'Wobbuffet') return 'Custap Berry';
-		if (species.name === 'Shuckle') return 'Mental Herb';
-		if (ability === 'Harvest' || ability === 'Cheek Pouch') return 'Sitrus Berry';
-		if (species.name === 'Ditto') return 'Choice Scarf';
-		if (ability === 'Poison Heal') return 'Toxic Orb';
-		if (ability === 'Speed Boost') return 'Life Orb';
-		if (species.nfe) return (species.name === 'Scyther' && role === 'Fast Attacker') ? 'Choice Band' : 'Eviolite';
+		if (species.id === 'farfetchd') return 'Stick';
+		if (species.id === 'latias' || species.id === 'latios') return 'Soul Dew';
+		if (species.id === 'marowak') return 'Thick Club';
+		if (species.id === 'pikachu') return 'Light Ball';
+		if (species.id === 'shedinja' || species.id === 'smeargle') return 'Focus Sash';
+		if (species.id === 'unown') return 'Choice Specs';
+		if (species.id === 'wobbuffet') return 'Custap Berry';
+		if (ability === 'Harvest') return 'Sitrus Berry';
+		if (species.id === 'ditto') return 'Choice Scarf';
+		if (species.id === 'exploud' && role === 'Bulky Attacker') return 'Choice Band';
+		if (ability === 'Poison Heal' || moves.has('facade')) return 'Toxic Orb';
+		if (ability === 'Speed Boost' && species.id !== 'ninjask') return 'Life Orb';
+		if (species.nfe) return 'Eviolite';
 		if (['healingwish', 'memento', 'switcheroo', 'trick'].some(m => moves.has(m))) {
-			if (species.baseStats.spe >= 60 && species.baseStats.spe <= 108 && role !== 'Wallbreaker') {
+			if (
+				species.baseStats.spe >= 60 && species.baseStats.spe <= 108 && role !== 'Wallbreaker' && !counter.get('priority')
+			) {
 				return 'Choice Scarf';
 			} else {
 				return (counter.get('Physical') > counter.get('Special')) ? 'Choice Band' : 'Choice Specs';
@@ -756,24 +700,18 @@ export class RandomGen6Teams extends RandomGen7Teams {
 		}
 		if (moves.has('bellydrum')) return 'Sitrus Berry';
 		if (moves.has('waterspout')) return 'Choice Scarf';
-		if (moves.has('geomancy') || moves.has('skyattack')) return 'Power Herb';
-		if (moves.has('shellsmash')) {
-			return (ability === 'Solid Rock' && !!counter.get('priority')) ? 'Weakness Policy' : 'White Herb';
-		}
+		if (moves.has('shellsmash')) return 'White Herb';
 		if (moves.has('psychoshift')) return 'Flame Orb';
-		if ((ability === 'Guts' || moves.has('facade')) && !moves.has('sleeptalk')) {
-			return species.name === 'Conkeldurr' ? 'Flame Orb' : 'Toxic Orb';
-		}
 		if (ability === 'Magic Guard' && role !== 'Bulky Support') {
 			return moves.has('counter') ? 'Focus Sash' : 'Life Orb';
 		}
 		if (species.id === 'rampardos' && role === 'Fast Attacker') return 'Choice Scarf';
 		if (ability === 'Sheer Force' && counter.get('sheerforce')) return 'Life Orb';
-		if (ability === 'Unburden') return (species.id === 'hitmonlee') ? 'White Herb' : 'Sitrus Berry';
-		if (moves.has('acrobatics')) return '';
+		if (moves.has('acrobatics')) return 'Flying Gem';
+		if (species.id === 'hitmonlee' && ability === 'Unburden') return moves.has('fakeout') ? 'Normal Gem' : 'Fighting Gem';
 		if (moves.has('lightscreen') && moves.has('reflect')) return 'Light Clay';
-		if (moves.has('rest') && !moves.has('sleeptalk') && !['Hydration', 'Natural Cure', 'Shed Skin'].includes(ability)) {
-			return 'Chesto Berry';
+		if (moves.has('rest') && !moves.has('sleeptalk') && !['Natural Cure', 'Shed Skin'].includes(ability)) {
+			return (moves.has('raindance') && ability === 'Hydration') ? 'Damp Rock' : 'Chesto Berry';
 		}
 		if (role === 'Staller') return 'Leftovers';
 	}
@@ -798,7 +736,8 @@ export class RandomGen6Teams extends RandomGen7Teams {
 		);
 
 		if (
-			moves.has('pursuit') && moves.has('suckerpunch') && counter.get('Dark') && !this.priorityPokemon.includes(species.id)
+			moves.has('pursuit') && moves.has('suckerpunch') && counter.get('Dark') &&
+			(!this.priorityPokemon.includes(species.id) || counter.get('Dark') >= 2)
 		) return 'Black Glasses';
 		if (counter.get('Special') === 4) {
 			return (
@@ -806,8 +745,8 @@ export class RandomGen6Teams extends RandomGen7Teams {
 			) ? 'Choice Scarf' : 'Choice Specs';
 		}
 		if (counter.get('Special') === 3 && moves.has('uturn')) return 'Choice Specs';
-		if (counter.get('Physical') === 4 && species.id !== 'jirachi' &&
-			['dragontail', 'fakeout', 'flamecharge', 'nuzzle', 'rapidspin'].every(m => !moves.has(m))
+		if (counter.get('Physical') === 4 && species.id !== 'jirachi' && species.id !== 'spinda' &&
+			['dragontail', 'fakeout', 'rapidspin'].every(m => !moves.has(m))
 		) {
 			return (
 				scarfReqs && (species.baseStats.atk >= 100 || ability === 'Pure Power' || ability === 'Huge Power') &&
@@ -815,24 +754,15 @@ export class RandomGen6Teams extends RandomGen7Teams {
 			) ? 'Choice Scarf' : 'Choice Band';
 		}
 
-		if (ability === 'Sturdy' && moves.has('explosion') && !counter.get('speedsetup')) return 'Custap Berry';
+		if (ability === 'Sturdy' && moves.has('explosion')) return 'Custap Berry';
 		if (types.includes('Normal') && moves.has('fakeout') && !!counter.get('Normal')) return 'Silk Scarf';
-		if (role === 'Bulky Setup' && !!counter.get('speedsetup') && !moves.has('swordsdance')) {
-			return 'Weakness Policy';
-		}
 		if (species.id === 'palkia') return 'Lustrous Orb';
-		if (species.id === 'archeops') return 'Expert Belt';
-		if (!counter.get('Status') && (
-			['Fast Support', 'Bulky Support', 'Bulky Attacker'].some(m => role === m) || moves.has('rapidspin')
-		)) {
-			return 'Assault Vest';
-		}
 		if (moves.has('outrage') && counter.get('setup')) return 'Lum Berry';
 		if (
-			(ability === 'Rough Skin') || (species.id !== 'hooh' &&
+			(ability === 'Rough Skin') || (species.id !== 'hooh' && role !== 'Wallbreaker' &&
 			ability === 'Regenerator' && species.baseStats.hp + species.baseStats.def >= 180 && this.randomChance(1, 2))
 		) return 'Rocky Helmet';
-		if (['kingsshield', 'protect', 'spikyshield', 'substitute'].some(m => moves.has(m))) return 'Leftovers';
+		if (['protect', 'substitute'].some(m => moves.has(m))) return 'Leftovers';
 		if (
 			this.dex.getEffectiveness('Ground', species) >= 2 &&
 			ability !== 'Levitate'
@@ -840,24 +770,31 @@ export class RandomGen6Teams extends RandomGen7Teams {
 			return 'Air Balloon';
 		}
 		if (
-			(role === 'Fast Support' || moves.has('stickyweb')) && isLead && defensiveStatTotal < 255 &&
-			!counter.get('recovery') && !moves.has('defog') && (!counter.get('recoil') || ability === 'Rock Head') &&
-			ability !== 'Regenerator'
+			role === 'Fast Support' && isLead && defensiveStatTotal < 255 && !counter.get('recovery') &&
+			(!counter.get('recoil') || ability === 'Rock Head') &&
+			(counter.get('hazards') || !(moves.has('uturn') || moves.has('voltswitch')))
 		) return 'Focus Sash';
 
 		// Default Items
 		if (role === 'Fast Support') {
 			return (
 				counter.get('Physical') + counter.get('Special') >= 3 &&
-				['nuzzle', 'rapidspin', 'uturn', 'voltswitch'].every(m => !moves.has(m)) &&
+				['rapidspin', 'uturn', 'voltswitch'].every(m => !moves.has(m)) &&
 				this.dex.getEffectiveness('Rock', species) < 2
 			) ? 'Life Orb' : 'Leftovers';
 		}
-		if (!counter.get('Status')) {
-			return (
-				(moves.has('uturn') || moves.has('voltswitch')) && !counter.get('Dragon') && !counter.get('Normal')
-			) ? 'Expert Belt' : 'Life Orb';
-		}
+		// noStab moves that should reject Expert Belt
+		const noExpertBeltMoves = (
+			this.noStab.filter(moveid => ['Dragon', 'Normal', 'Poison'].includes(this.dex.moves.get(moveid).type))
+		);
+		const expertBeltReqs = (
+			!counter.get('Dragon') && !counter.get('Normal') && !counter.get('Poison') &&
+			noExpertBeltMoves.every(m => !moves.has(m))
+		);
+		if (
+			!counter.get('Status') && expertBeltReqs &&
+			(moves.has('uturn') || moves.has('voltswitch') || role === 'Fast Attacker')
+		) return 'Expert Belt';
 		if (
 			['Fast Attacker', 'Setup Sweeper', 'Wallbreaker'].some(m => role === m) &&
 			this.dex.getEffectiveness('Rock', species) < 2 && ability !== 'Sturdy'
@@ -874,7 +811,18 @@ export class RandomGen6Teams extends RandomGen7Teams {
 		const forme = this.getForme(species);
 		const sets = this.randomSets[species.id]["sets"];
 		const possibleSets = [];
-		for (const set of sets) possibleSets.push(set);
+		// Check if the Pokemon has a Spinner set
+		let canSpinner = false;
+		for (const set of sets) {
+			if (!teamDetails.rapidSpin && set.role === 'Spinner') canSpinner = true;
+		}
+		for (const set of sets) {
+			// Prevent Spinner if the team already has removal
+			if (teamDetails.rapidSpin && set.role === 'Spinner') continue;
+			// Enforce Spinner if the team does not have removal
+			if (canSpinner && set.role !== 'Spinner') continue;
+			possibleSets.push(set);
+		}
 		const set = this.sampleIfArray(possibleSets);
 		const role = set.role;
 		const movePool: string[] = Array.from(set.movepool);
@@ -913,15 +861,6 @@ export class RandomGen6Teams extends RandomGen7Teams {
 
 		const level = this.getLevel(species);
 
-		// Minimize confusion damage, including if Foul Play is its only physical attack
-		if (
-			(!counter.get('Physical') || (counter.get('Physical') <= 1 && moves.has('foulplay'))) &&
-			!moves.has('copycat') && !moves.has('transform')
-		) {
-			evs.atk = 0;
-			ivs.atk = 0;
-		}
-
 		// We use a special variable to track Hidden Power
 		// so that we can check for all Hidden Powers at once
 		let hasHiddenPower = false;
@@ -935,7 +874,7 @@ export class RandomGen6Teams extends RandomGen7Teams {
 				if (move.startsWith('hiddenpower')) hpType = move.substr(11);
 			}
 			if (!hpType) throw new Error(`hasHiddenPower is true, but no Hidden Power move was found.`);
-			const HPivs = ivs.atk === 0 ? ZeroAttackHPIVs[hpType] : this.dex.types.get(hpType).HPivs;
+			const HPivs = this.dex.types.get(hpType).HPivs;
 			let iv: StatID;
 			for (iv in HPivs) {
 				ivs[iv] = HPivs[iv]!;
@@ -970,9 +909,15 @@ export class RandomGen6Teams extends RandomGen7Teams {
 			evs.hp -= 4;
 		}
 
+		// Minimize confusion damage, including if Foul Play is its only physical attack
+		if ((!counter.get('Physical') || (counter.get('Physical') <= 1 && moves.has('foulplay'))) && !moves.has('transform')) {
+			evs.atk = 0;
+			ivs.atk = hasHiddenPower ? (ivs.atk || 31) - 28 : 0;
+		}
+
 		if (['gyroball', 'metalburst', 'trickroom'].some(m => moves.has(m))) {
 			evs.spe = 0;
-			ivs.spe = (hasHiddenPower && level < 100) ? ivs.spe - 30 : 0;
+			ivs.spe = hasHiddenPower ? (ivs.spe || 31) - 28 : 0;
 		}
 
 		// shuffle moves to add more randomness to camomons
@@ -994,237 +939,132 @@ export class RandomGen6Teams extends RandomGen7Teams {
 		};
 	}
 
-	randomFactorySets: {[format: string]: {[species: string]: BattleFactorySpecies}} = require('./factory-sets.json');
-
-	randomFactorySet(
-		species: Species,
-		teamData: RandomTeamsTypes.FactoryTeamDetails,
-		tier: string
-	): RandomTeamsTypes.RandomFactorySet | null {
-		const id = toID(species.name);
-		// const flags = this.randomFactorySets[tier][id].flags;
-		const setList = this.randomFactorySets[tier][id].sets;
-
-		const itemsMax: {[k: string]: number} = {choicespecs: 1, choiceband: 1, choicescarf: 1};
-		const movesMax: {[k: string]: number} = {
-			rapidspin: 1, batonpass: 1, stealthrock: 1, defog: 1, spikes: 1, toxicspikes: 1,
-		};
-		const requiredMoves: {[k: string]: string} = {stealthrock: 'hazardSet', rapidspin: 'hazardClear', defog: 'hazardClear'};
-		const weatherAbilitiesRequire: {[k: string]: string} = {
-			hydration: 'raindance', swiftswim: 'raindance',
-			leafguard: 'sunnyday', solarpower: 'sunnyday', chlorophyll: 'sunnyday',
-			sandforce: 'sandstorm', sandrush: 'sandstorm', sandveil: 'sandstorm',
-			snowcloak: 'hail',
-		};
-		const weatherAbilities = ['drizzle', 'drought', 'snowwarning', 'sandstream'];
-
-		// Build a pool of eligible sets, given the team partners
-		// Also keep track of sets with moves the team requires
-		let effectivePool: {set: AnyObject, moveVariants?: number[], itemVariants?: number, abilityVariants?: number}[] = [];
-		const priorityPool = [];
-		for (const curSet of setList) {
-			if (this.forceMonotype && !species.types.includes(this.forceMonotype)) continue;
-
-			const itemData = this.dex.items.get(curSet.item);
-			if (teamData.megaCount && teamData.megaCount > 0 && itemData.megaStone) continue; // reject 2+ mega stones
-			if (itemsMax[itemData.id] && teamData.has[itemData.id] >= itemsMax[itemData.id]) continue;
-
-			const abilityState = this.dex.abilities.get(curSet.ability);
-			if (weatherAbilitiesRequire[abilityState.id] && teamData.weather !== weatherAbilitiesRequire[abilityState.id]) continue;
-			if (teamData.weather && weatherAbilities.includes(abilityState.id)) continue; // reject 2+ weather setters
-
-			let reject = false;
-			let hasRequiredMove = false;
-			const curSetVariants = [];
-			for (const move of curSet.moves) {
-				const variantIndex = this.random(move.length);
-				const moveId = toID(move[variantIndex]);
-				if (movesMax[moveId] && teamData.has[moveId] >= movesMax[moveId]) {
-					reject = true;
-					break;
-				}
-				if (requiredMoves[moveId] && !teamData.has[requiredMoves[moveId]]) {
-					hasRequiredMove = true;
-				}
-				curSetVariants.push(variantIndex);
-			}
-			if (reject) continue;
-			effectivePool.push({set: curSet, moveVariants: curSetVariants});
-			if (hasRequiredMove) priorityPool.push({set: curSet, moveVariants: curSetVariants});
-		}
-		if (priorityPool.length) effectivePool = priorityPool;
-
-		if (!effectivePool.length) {
-			if (!teamData.forceResult) return null;
-			for (const curSet of setList) {
-				effectivePool.push({set: curSet});
-			}
-		}
-
-		const setData = this.sample(effectivePool);
-		const moves = [];
-		for (const [i, moveSlot] of setData.set.moves.entries()) {
-			moves.push(setData.moveVariants ? moveSlot[setData.moveVariants[i]] : this.sample(moveSlot));
-		}
-
-		return {
-			name: setData.set.name || species.baseSpecies,
-			species: setData.set.species,
-			gender: setData.set.gender || species.gender || (this.randomChance(1, 2) ? 'M' : 'F'),
-			item: setData.set.item || '',
-			ability: setData.set.ability || species.abilities['0'],
-			shiny: typeof setData.set.shiny === 'undefined' ? this.randomChance(1, 1024) : setData.set.shiny,
-			level: this.adjustLevel || 100,
-			happiness: typeof setData.set.happiness === 'undefined' ? 255 : setData.set.happiness,
-			evs: setData.set.evs || {hp: 84, atk: 84, def: 84, spa: 84, spd: 84, spe: 84},
-			ivs: setData.set.ivs || {hp: 31, atk: 31, def: 31, spa: 31, spd: 31, spe: 31},
-			nature: setData.set.nature || 'Serious',
-			moves: moves,
-		};
-	}
-
-	randomFactoryTeam(side: PlayerOptions, depth = 0): RandomTeamsTypes.RandomFactorySet[] {
+	randomTeam() {
 		this.enforceNoDirectCustomBanlistChanges();
 
-		const forceResult = (depth >= 12);
+		const seed = this.prng.seed;
+		const ruleTable = this.dex.formats.getRuleTable(this.format);
+		const pokemon: RandomTeamsTypes.RandomSet[] = [];
 
-		// The teams generated depend on the tier choice in such a way that
-		// no exploitable information is leaked from rolling the tier in getTeam(p1).
-		if (!this.factoryTier) this.factoryTier = this.sample(['Uber', 'OU', 'UU', 'RU', 'NU', 'PU']);
-		const chosenTier = this.factoryTier;
+		// For Monotype
+		const isMonotype = !!this.forceMonotype || ruleTable.has('sametypeclause');
+		const typePool = this.dex.types.names();
+		const type = this.forceMonotype || this.sample(typePool);
 
-		const pokemon = [];
+		const baseFormes: {[k: string]: number} = {};
+		const typeCount: {[k: string]: number} = {};
+		const typeWeaknesses: {[k: string]: number} = {};
+		const typeDoubleWeaknesses: {[k: string]: number} = {};
+		const teamDetails: RandomTeamsTypes.TeamDetails = {};
+		let numMaxLevelPokemon = 0;
 
-		const pokemonPool = Object.keys(this.randomFactorySets[chosenTier]);
-
-		const teamData: TeamData = {
-			typeCount: {}, typeComboCount: {}, baseFormes: {}, megaCount: 0, has: {}, forceResult,
-			weaknesses: {}, resistances: {},
-		};
-		const requiredMoveFamilies = ['hazardSet', 'hazardClear'];
-		const requiredMoves: {[k: string]: string} = {stealthrock: 'hazardSet', rapidspin: 'hazardClear', defog: 'hazardClear'};
-		const weatherAbilitiesSet: {[k: string]: string} = {
-			drizzle: 'raindance', drought: 'sunnyday', snowwarning: 'hail', sandstream: 'sandstorm',
-		};
-		const resistanceAbilities: {[k: string]: string[]} = {
-			dryskin: ['Water'], waterabsorb: ['Water'], stormdrain: ['Water'],
-			flashfire: ['Fire'], heatproof: ['Fire'],
-			lightningrod: ['Electric'], motordrive: ['Electric'], voltabsorb: ['Electric'],
-			sapsipper: ['Grass'],
-			thickfat: ['Ice', 'Fire'],
-			levitate: ['Ground'],
-		};
-
-		while (pokemonPool.length && pokemon.length < this.maxTeamSize) {
-			const species = this.dex.species.get(this.sampleNoReplace(pokemonPool));
+		const pokemonList = Object.keys(this.randomSets);
+		const [pokemonPool, baseSpeciesPool] = this.getPokemonPool(type, pokemon, isMonotype, pokemonList);
+		while (baseSpeciesPool.length && pokemon.length < this.maxTeamSize) {
+			const baseSpecies = this.sampleNoReplace(baseSpeciesPool);
+			const species = this.dex.species.get(this.sample(pokemonPool[baseSpecies]));
 			if (!species.exists) continue;
 
-			const speciesFlags = this.randomFactorySets[chosenTier][species.id].flags;
-
 			// Limit to one of each species (Species Clause)
-			if (teamData.baseFormes[species.baseSpecies]) continue;
+			if (baseFormes[species.baseSpecies]) continue;
 
-			// Limit the number of Megas to one
-			if (!teamData.megaCount) teamData.megaCount = 0;
-			if (teamData.megaCount >= 1 && speciesFlags.megaOnly) continue;
+			// Illusion shouldn't be in the last slot
+			if (species.name === 'Zoroark' && pokemon.length >= (this.maxTeamSize - 1)) continue;
 
 			// Dynamically scale limits for different team sizes. The default and minimum value is 1.
 			const limitFactor = Math.round(this.maxTeamSize / 6) || 1;
 
-			// Limit 2 of any type
 			const types = species.types;
-			let skip = false;
-			for (const type of types) {
-				if (teamData.typeCount[type] >= 2 * limitFactor && this.randomChance(4, 5)) {
-					skip = true;
-					break;
+
+			if (!isMonotype && !this.forceMonotype) {
+				let skip = false;
+
+				// Limit two of any type
+				for (const typeName of types) {
+					if (typeCount[typeName] >= 2 * limitFactor) {
+						skip = true;
+						break;
+					}
+				}
+				if (skip) continue;
+
+				// Limit three weak to any type, and one double weak to any type
+				for (const typeName of this.dex.types.names()) {
+					// it's weak to the type
+					if (this.dex.getEffectiveness(typeName, species) > 0) {
+						if (!typeWeaknesses[typeName]) typeWeaknesses[typeName] = 0;
+						if (typeWeaknesses[typeName] >= 3 * limitFactor) {
+							skip = true;
+							break;
+						}
+					}
+					if (this.dex.getEffectiveness(typeName, species) > 1) {
+						if (!typeDoubleWeaknesses[typeName]) typeDoubleWeaknesses[typeName] = 0;
+						if (typeDoubleWeaknesses[typeName] >= 1 * limitFactor) {
+							skip = true;
+							break;
+						}
+					}
+				}
+				if (skip) continue;
+
+				// Limit one level 100 Pokemon
+				if (!this.adjustLevel && (this.getLevel(species) === 100) && numMaxLevelPokemon >= limitFactor) {
+					continue;
 				}
 			}
-			if (skip) continue;
 
-			const set = this.randomFactorySet(species, teamData, chosenTier);
-			if (!set) continue;
-
-			// Limit 1 of any type combination
-			let typeCombo = types.slice().sort().join();
-			if (set.ability === 'Drought' || set.ability === 'Drizzle') {
-				// Drought and Drizzle don't count towards the type combo limit
-				typeCombo = set.ability;
-			}
-			if (teamData.typeComboCount[typeCombo] >= 1 * limitFactor) continue;
+			const set = this.randomSet(species, teamDetails, pokemon.length === 0);
 
 			// Okay, the set passes, add it to our team
 			pokemon.push(set);
 
-			// Now that our Pokemon has passed all checks, we can update team data:
-			for (const type of types) {
-				if (type in teamData.typeCount) {
-					teamData.typeCount[type]++;
+			// Don't bother tracking details for the last Pokemon
+			if (pokemon.length === this.maxTeamSize) break;
+
+			// Now that our Pokemon has passed all checks, we can increment our counters
+			baseFormes[species.baseSpecies] = 1;
+
+			// Increment type counters
+			for (const typeName of types) {
+				if (typeName in typeCount) {
+					typeCount[typeName]++;
 				} else {
-					teamData.typeCount[type] = 1;
-				}
-			}
-			teamData.typeComboCount[typeCombo] = (teamData.typeComboCount[typeCombo] + 1) || 1;
-
-			teamData.baseFormes[species.baseSpecies] = 1;
-
-			const itemData = this.dex.items.get(set.item);
-			if (itemData.megaStone) teamData.megaCount++;
-			if (itemData.id in teamData.has) {
-				teamData.has[itemData.id]++;
-			} else {
-				teamData.has[itemData.id] = 1;
-			}
-
-			const abilityState = this.dex.abilities.get(set.ability);
-			if (abilityState.id in weatherAbilitiesSet) {
-				teamData.weather = weatherAbilitiesSet[abilityState.id];
-			}
-
-			for (const move of set.moves) {
-				const moveId = toID(move);
-				if (moveId in teamData.has) {
-					teamData.has[moveId]++;
-				} else {
-					teamData.has[moveId] = 1;
-				}
-				if (moveId in requiredMoves) {
-					teamData.has[requiredMoves[moveId]] = 1;
+					typeCount[typeName] = 1;
 				}
 			}
 
+			// Increment weakness counter
 			for (const typeName of this.dex.types.names()) {
-				// Cover any major weakness (3+) with at least one resistance
-				if (teamData.resistances[typeName] >= 1) continue;
-				if (resistanceAbilities[abilityState.id]?.includes(typeName) || !this.dex.getImmunity(typeName, types)) {
-					// Heuristic: assume that Pokemon with these abilities don't have (too) negative typing.
-					teamData.resistances[typeName] = (teamData.resistances[typeName] || 0) + 1;
-					if (teamData.resistances[typeName] >= 1) teamData.weaknesses[typeName] = 0;
-					continue;
+				// it's weak to the type
+				if (this.dex.getEffectiveness(typeName, species) > 0) {
+					typeWeaknesses[typeName]++;
 				}
-				const typeMod = this.dex.getEffectiveness(typeName, types);
-				if (typeMod < 0) {
-					teamData.resistances[typeName] = (teamData.resistances[typeName] || 0) + 1;
-					if (teamData.resistances[typeName] >= 1) teamData.weaknesses[typeName] = 0;
-				} else if (typeMod > 0) {
-					teamData.weaknesses[typeName] = (teamData.weaknesses[typeName] || 0) + 1;
+				if (this.dex.getEffectiveness(typeName, species) > 1) {
+					typeDoubleWeaknesses[typeName]++;
 				}
 			}
-		}
-		if (pokemon.length < this.maxTeamSize) return this.randomFactoryTeam(side, ++depth);
 
-		// Quality control
-		if (!teamData.forceResult) {
-			for (const requiredFamily of requiredMoveFamilies) {
-				if (!teamData.has[requiredFamily]) return this.randomFactoryTeam(side, ++depth);
-			}
-			for (const type in teamData.weaknesses) {
-				if (teamData.weaknesses[type] >= 3) return this.randomFactoryTeam(side, ++depth);
-			}
+			// Increment level 100 counter
+			if (set.level === 100) numMaxLevelPokemon++;
+
+			// Team details
+			if (set.ability === 'Snow Warning' || set.moves.includes('hail')) teamDetails.hail = 1;
+			if (set.ability === 'Drizzle' || set.moves.includes('raindance')) teamDetails.rain = 1;
+			if (set.ability === 'Sand Stream') teamDetails.sand = 1;
+			if (set.ability === 'Drought' || set.moves.includes('sunnyday')) teamDetails.sun = 1;
+			if (set.moves.includes('spikes')) teamDetails.spikes = (teamDetails.spikes || 0) + 1;
+			if (set.moves.includes('stealthrock')) teamDetails.stealthRock = 1;
+			if (set.moves.includes('toxicspikes')) teamDetails.toxicSpikes = 1;
+			if (set.moves.includes('rapidspin')) teamDetails.rapidSpin = 1;
+			if (set.moves.includes('reflect') && set.moves.includes('lightscreen')) teamDetails.screens = 1;
+		}
+		if (pokemon.length < this.maxTeamSize && pokemon.length < 12) {
+			throw new Error(`Could not build a random team for ${this.format} (seed=${seed})`);
 		}
 
 		return pokemon;
 	}
 }
 
-export default RandomGen6Teams;
+export default RandomGen5Teams;
