+++ data/random-battles/gen6-sets.json
