+++ data/random-battles/gen7-factory-sets.json
