+++ data/random-battles/gen7-sets.json
