+++ data/random-battles/gen8-data.json
