+++ data/random-battles/gen8bdsp-data.json
