+++ data/random-battles/gen9-doubles-sets.json
